refactor: phase out session, transfer and parts of fs_plan into transfer_flow
