+++ V0.2.0_SCOPE.md
@@ -1,0 +1,24 @@
+# Drift v0.2.0 Scope
+
+## 1. Improved Transfer Cancellation
+- [x] **Protocol Design:** Implement dedicated control messages for graceful user-initiated aborts.
+- [x] **Reliability:** Ensure peers are immediately notified of cancellations to prevent hangs.
+- [x] **Verification:** Comprehensive test coverage for cancellation flows (sender-side and receiver-side).
+
+## 2. Granular Progress Tracking
+- [ ] **Per-file Progress:** Track and report individual file transfer status.
+- [ ] **Metrics:** Calculate and expose real-time transfer speed (e.g., sliding window).
+- [ ] **Feedback:** Provide Estimated Time of Arrival (ETA) based on current speed.
+
+## 3. Advanced Error Handling & Diagnostics
+- [ ] **Structured Errors:** Move from generic strings to a structured `DriftError` enum.
+- [ ] **Actionable Diagnostics:** Provide specific advice for common network issues (NAT, firewalls).
+- [ ] **Metadata:** Show connection paths (e.g., "Direct" vs "Relay") to help debug performance.
+
+## 4. Enhanced Conflict Management
+- [x] **ConflictPolicy::Rename:** Add a new conflict resolution strategy.
+- [x] **Implementation:** Automatically rename conflicting files (e.g., `a.pdf` becomes `a (1).pdf`) instead of just `Reject` or `Overwrite`.
+- [x] **Consistency:** Ensure this policy is available across both CLI and Flutter app.
+
+---
+*Note: This file is for tracking the planned v0.2.0 scope and should not be committed to source control.*
