+++ crates/core/src/lib.rs
@@ -3,8 +3,6 @@ pub mod fs_plan;
 pub mod lan;
 pub mod pairing;
 pub mod rendezvous;
-pub mod session;
-pub mod transfer;
 pub mod util;
 
 pub mod blobs;
